Extract operation dispatcher
